+++ Physics3D_class5 - solution/Timer.h
@@ -14,7 +14,7 @@ public:
 	void Start();
 	void Stop();
 
-	Uint32 Read();
+	float Read();
 
 private:
